Software-Engineering-1402-02/src/requirements.txt
Django==4.1.3
djangorestframework
pyyaml
requests
django-cors-headers
pillow
django-storages
boto3
mysql-connector-python